@@ -1,7 +1,12 @@
+import 'dart:async';
 import 'dart:io';
 
+import 'package:hive_ce/hive.dart';
+import 'package:collection/collection.dart';
 import 'package:injectable/injectable.dart';
-import 'package:isar/isar.dart';
+import 'package:logging/logging.dart';
+import 'package:synchronized/synchronized.dart';
+import 'package:uuid/uuid.dart';
 
 import '../../domain/exceptions/exceptions.dart';
 import '../../domain/models/auth_model.dart';
@@ -12,122 +17,183 @@ import '../entities/entities.dart';
 
 @Singleton(as: DataBaseRepository)
 class DataBaseRepositoryImpl implements DataBaseRepository {
+  final _log = Logger('DataBaseRepositoryImpl');
   bool _isInneted = false;
-  late final Isar _db;
+  Timer? _timer;
+  final _boxOpenedTime = Map<String, DateTime>.identity();
+  final _lock = Lock(reentrant: true);
 
   @override
   @PostConstruct()
-  Future<void> init() {
+  void init() {
     final directory = Platform.environment['DB_DIRECTORY'] ?? 'db';
-    return Isar.openAsync(
-      directory: directory,
-      schemas: [
-        EmailAuthEntitySchema,
-        OpenAuthEntitySchema,
-        ProfileEntitySchema,
-        SessionEntitySchema,
-      ],
-    ).then((value) {
-      _db = value;
-      _isInneted = true;
-    });
+    final dbUnUsedClosePeriod =
+        int.tryParse(Platform.environment['DB_UNUSED_CLOSE_PERIOD'] ?? '') ??
+            10;
+
+    _log.config('DB_DIRECTORY is $directory');
+    _log.config('DB_UNUSED_CLOSE_PERIOD is $dbUnUsedClosePeriod');
+    if (!_isInneted) {
+      Hive.init(directory);
+    }
+
+    final openAuthEntityAdapter = OpenAuthEntityAdapter();
+    final emailAuthEntityAdapter = EmailAuthEntityAdapter();
+    final profileEntityAdapter = ProfileEntityAdapter();
+    final sessionEntityAdapter = SessionEntityAdapter();
+
+    if (!Hive.isAdapterRegistered(openAuthEntityAdapter.typeId)) {
+      Hive.registerAdapter(openAuthEntityAdapter);
+    }
+    if (!Hive.isAdapterRegistered(emailAuthEntityAdapter.typeId)) {
+      Hive.registerAdapter(emailAuthEntityAdapter);
+    }
+    if (!Hive.isAdapterRegistered(profileEntityAdapter.typeId)) {
+      Hive.registerAdapter(profileEntityAdapter);
+    }
+    if (!Hive.isAdapterRegistered(sessionEntityAdapter.typeId)) {
+      Hive.registerAdapter(sessionEntityAdapter);
+    }
+
+    if (_timer?.isActive == true) {
+      _timer?.cancel();
+    }
+
+    Timer.periodic(
+        Duration(seconds: dbUnUsedClosePeriod), (_) => _closeUnusedBoxs());
+    _isInneted = true;
   }
 
   @override
-  Future<ProfileModel?> getProfile(int id) => // TODO: implement
-      throw UnimplementedError() /* Future.value(_db.box<ProfileEntity>().get(id)?.toModel()) */;
+  Future<ProfileModel?> getProfile(String id) => _getBox<ProfileEntity>()
+      .then((box) => box.get(id))
+      .then((value) => value?.toModel());
 
   @override
-  Future<ProfileModel> putProfile(ProfileModel profile) => // TODO: implement
-      throw UnimplementedError() /* Future.value(_db.box<ProfileEntity>().put(ProfileEntity.fromModel(profile)))
-      .then((value) => profile.rebuild((p0) => p0..id = value)
-      ) */
-      ;
-
-  @override
-  Future<void> deleteProfile(int id) => // TODO: implement
-      throw UnimplementedError() /* Future.value( _db.box<ProfileEntity>().remove(id)).then((_) {}) */;
-
-  @override
-  Future<SessionModel?> getSession(int id) => // TODO: implement
-      throw UnimplementedError() /* Future.value(_db.box<SessionEntity>().get(id)?.toModel()) */;
-
-  @override
-  Future<List<SessionModel>> getSessionsBy(int profileId, String deviceId) {
-    // TODO: implement
-    throw UnimplementedError()
-        /* QueryBuilder<SessionEntity> query = _db
-        .box<SessionEntity>()
-        .query(SessionEntity_.deviceId.equals(deviceId))
-      ..link(SessionEntity_.user, ProfileEntity_.id.equals(profileId));
-    return Future.value(query.build().find())
-        .then((value) => value.map((e) => e.toModel()).toList()
-        ) */
-        ;
+  Future<ProfileModel> putProfile(ProfileModel profile) {
+    final id = profile.id ?? Uuid().v1();
+    return _getBox<ProfileEntity>()
+        .then((box) => box.put(id, ProfileEntity.fromModel(profile)))
+        .then((_) => profile.rebuild((p0) => p0..id = id));
   }
 
   @override
-  Future<List<SessionModel>> getExpiredSessions() => // TODO: implement
-      throw UnimplementedError() /* Future.value(_db
-          .box<SessionEntity>()
-          .query(SessionEntity_.dateExpired
-              .lessThan(DateTime.now().toUtc().millisecondsSinceEpoch))
-          .build()
-          .find())
-      .then((value) => value.map((e) => e.toModel()).toList()) */
-      ;
+  Future<void> deleteProfile(String id) =>
+      _getBox<ProfileEntity>().then((box) => box.delete(id));
 
   @override
-  Future<SessionModel> putSession(SessionModel session) => // TODO: implement
-      throw UnimplementedError() /* Future.value(_db.box<SessionEntity>().put(SessionEntity.fromModel(session)))
-      .then((value) => session.rebuild((p0) => p0..id = value)) */
-      ;
+  Future<SessionModel?> getSession(String id) => _getBox<SessionEntity>()
+      .then((box) => box.get(id))
+      .then((value) => value?.toModel());
 
   @override
-  Future<void> deleteSession(int id) => // TODO: implement
-      throw UnimplementedError() /* Future.value(_db.box<SessionEntity>().remove(id)).then((_) {}) */;
+  Future<List<SessionModel>> getSessionsBy(String profileId, String deviceId) =>
+      _getBox<SessionEntity>()
+          .then((box) => box.values.where((element) =>
+              element.userId == profileId && element.deviceId == deviceId))
+          .then((value) => value.map((e) => e.toModel()).toList());
 
   @override
-  Future<void> deleteSessions(List<int> ids) => // TODO: implement
-      throw UnimplementedError() /* Future.value(_db.box<SessionEntity>().removeMany(ids)).then((_) {}) */;
+  Future<List<SessionModel>> getExpiredSessions() => _getBox<SessionEntity>()
+      .then((box) => box.values.where(
+          (element) => element.dateExpired.isAfter(DateTime.now().toUtc())))
+      .then((value) => value.map((e) => e.toModel()).toList());
 
   @override
-  Future<EmailAuthModel?> getAuthByEmail(String email) => // TODO: implement
-      throw UnimplementedError() /* Future.value(_db
-      .box<EmailAuthEntity>()
-      .query(EmailAuthEntity_.email.equals(email))
-      .build()
-      .findFirst()
-      ?.toModel()) */
-      ;
+  Future<SessionModel> putSession(SessionModel session) {
+    final id = session.id ?? Uuid().v1();
+    return _getBox<SessionEntity>()
+        .then((box) => box.put(id, SessionEntity.fromModel(session)))
+        .then((_) => session.rebuild((p0) => p0..id = id));
+  }
 
   @override
-  Future<void> deleteEmailAuth(int id) => // TODO: implement
-      throw UnimplementedError() /* Future.value(_db.box<EmailAuthEntity>().remove(id)).then((_) {}) */;
+  Future<void> deleteSession(String id) =>
+      _getBox<SessionEntity>().then((box) => box.delete(id));
 
   @override
-  Future<EmailAuthModel?> getEmailAuth(int id) => // TODO: implement
-      throw UnimplementedError() /*Future.value(_db.box<EmailAuthEntity>().get(id)?.toModel())*/;
+  Future<void> deleteSessions(List<String> ids) =>
+      _getBox<SessionEntity>().then((box) => box.deleteAll(ids));
 
   @override
-  Future<EmailAuthModel> putEmailAuth(EmailAuthModel auth) =>
-      // TODO: implement
-      throw UnimplementedError() /* Future.value(_db.box<EmailAuthEntity>().put(EmailAuthEntity.fromModel(auth)))
-      .then((value) => auth.rebuild((p0) => p0..id = value) 
-      )*/
-      ;
+  Future<EmailAuthModel?> getAuthByEmail(
+          String email) =>
+      _getBox<EmailAuthEntity>()
+          .then((box) =>
+              box.values.firstWhereOrNull((element) => element.email == email))
+          .then((value) => value?.toModel());
+
+  @override
+  Future<void> deleteEmailAuth(String id) =>
+      _getBox<EmailAuthEntity>().then((box) => box.delete(id));
+
+  @override
+  Future<EmailAuthModel?> getEmailAuth(String id) => _getBox<EmailAuthEntity>()
+      .then((box) => box.get(id))
+      .then((value) => value?.toModel());
+
+  @override
+  Future<EmailAuthModel> putEmailAuth(EmailAuthModel auth) {
+    final id = auth.id ?? Uuid().v1();
+    return _getBox<EmailAuthEntity>()
+        .then((box) => box.put(id, EmailAuthEntity.fromModel(auth)))
+        .then((_) => auth.rebuild((p0) => p0..id = id));
+  }
 
   @override
   @disposeMethod
-  void dispose() {
-    _getDb.close();
+  Future<void> dispose() async {
+    _timer?.cancel();
+    await Hive.close().then((_) {
+      _isInneted = false;
+    });
   }
 
   @override
   bool get isInneted => _isInneted;
 
-  Isar get _getDb {
+  void _closeUnusedBoxs() {
+    _lock.synchronized(() {
+      final boxes = _boxOpenedTime.entries
+          .where((element) => element.value.isAfter(DateTime.now().toUtc()));
+      for (var boxName in boxes.map((e) => e.key)) {
+        if (Hive.isBoxOpen(boxName)) {
+          final box = switch (boxName) {
+            'open_auth_box' => Hive.box<OpenAuthEntity>(boxName),
+            'email_auth_box' => Hive.box<EmailAuthEntity>(boxName),
+            'profile_box' => Hive.box<ProfileEntity>(boxName),
+            'session_box' => Hive.box<SessionEntity>(boxName),
+            _ => throw DbException.internalError()
+          };
+          return box.close().then((_) {
+            _log.fine('Close Unused $boxName Box');
+            _boxOpenedTime.removeWhere((key, value) => key == boxName);
+          });
+        }
+        _boxOpenedTime.removeWhere((key, value) => key == boxName);
+      }
+    });
+  }
+
+  Future<Box<T>> _getBox<T extends HiveObjectMixin>() {
     if (!isInneted) throw DbException.notInneted();
-    return _db;
+    return _lock.synchronized<Box<T>>(() {
+      final boxName = switch (T.toString()) {
+        'OpenAuthEntity' => 'open_auth_box',
+        'EmailAuthEntity' => 'email_auth_box',
+        'ProfileEntity' => 'profile_box',
+        'SessionEntity' => 'session_box',
+        _ => throw DbException.internalError()
+      };
+
+      _boxOpenedTime[boxName] = DateTime.now().toUtc();
+      if (Hive.isBoxOpen(boxName)) {
+        return Future.value(Hive.box<T>(boxName));
+      } else {
+        return Hive.openBox<T>(boxName).whenComplete(() {
+          _log.fine('Open $boxName Box');
+        });
+      }
+    });
   }
 }
