@@ -1,21 +1,25 @@
+// ignore_for_file: must_be_immutable
+
 import 'package:built_collection/built_collection.dart';
-import 'package:isar/isar.dart';
+import 'package:equatable/equatable.dart';
+import 'package:hive_ce/hive.dart';
 
 import '../../domain/models/models.dart';
 
 part 'profile_entity.g.dart';
 
-@collection
-class ProfileEntity {
-  @Id()
-  int id;
+@HiveType(typeId: 2)
+class ProfileEntity with HiveObjectMixin, EquatableMixin {
+  @HiveField(0)
   final String? name;
+  @HiveField(1)
   final double? price;
+  @HiveField(2)
   final List<int> workDays;
+  @HiveField(3)
   final DateTime dateCreated;
 
   ProfileEntity({
-    this.id = 0,
     this.name,
     this.price,
     this.workDays = const <int>[],
@@ -23,17 +27,19 @@ class ProfileEntity {
   });
 
   ProfileModel toModel() => ProfileModel((p0) => p0
-    ..id = id
+    ..id = key.toString()
     ..name = name
     ..price = price
     ..workDays = ListBuilder(workDays)
     ..dateCreated = dateCreated);
 
   factory ProfileEntity.fromModel(ProfileModel model) => ProfileEntity(
-        id: model.id ?? 0,
         name: model.name,
         price: model.price,
         workDays: model.workDays.toList(),
         dateCreated: model.dateCreated,
       );
+
+  @override
+  List<Object?> get props => [];
 }
