@@ -4,8 +4,8 @@ part 'session_model.g.dart';
 
 abstract class SessionModel
     implements Built<SessionModel, SessionModelBuilder> {
-  int? get id;
-  int get userId;
+  String? get id;
+  String get userId;
   String get deviceId;
   DateTime get dateCreated;
   DateTime get dateExpired;
